% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_example.R
\name{read_example}
\alias{read_example}
\title{Read example datasets shipped with the package}
\usage{
read_example(
  name = c("meling_grs_all.csv", "meling_grs_org.csv", "oyelade_sdnn_all.csv",
    "oyelade_sdnn_org.csv")
)
}
\arguments{
\item{name}{character, one of
\code{"oyelade_sdnn.csv"}, \code{"meling_grs.csv"},
\code{"oyelade_sdnn_org.csv"}, \code{"meling_grs_org.csv"}}
}
\value{
data.frame
}
\description{
Read example datasets shipped with the package
}
\examples{
read_example("oyelade_sdnn.csv")
}
